+++ src/components/position.rs
@@ -1,0 +1,8 @@
+use specs::VecStorage;
+
+#[derive(Component, Debug)]
+#[storage(VecStorage)]
+pub struct Position {
+    pub x: f32,
+    pub y: f32
+}
