+++ src/components/size.rs
@@ -1,0 +1,8 @@
+use specs::VecStorage;
+
+#[derive(Component, Debug)]
+#[storage(VecStorage)]
+pub struct Size {
+    pub width: i32,
+    pub height: i32
+}
